Move connection check mechanism + fix build #1935
